EDUX CI initial deployment (#3)
* Init deploy DW script

Initiating deployment script for dokuwiki. It should be used in the future to automatically transform and upload pages to Edux system from CI. Inspired by cvut/MI-PYT course.

* Init deploy DW script

Initiating deployment script for dokuwiki. It should be used in the future to automatically transform and upload pages to Edux system from CI. Inspired by cvut/MI-PYT course.

* Filters for enhancing transformation

Added filters to the app. One for interlinking local documents and second for Haskell syntax highlighting which was not provided by default (need to make use of Pascal there).

* Try to setup CI with envvars

* Simplify Travis config

* Try install pandoc from .deb

* Just master branch will be deployed